requirements.txt
kafka-python
apache-beam
schedule
mysql-connector-python